+++ src/Acquisition/__init__.py
@@ -3,7 +3,6 @@
 
 import copyreg
 import os
-import operator
 import platform
 import sys
 import types
@@ -17,13 +16,9 @@ from .interfaces import IAcquirer
 from .interfaces import IAcquisitionWrapper
 
 IS_PYPY = getattr(platform, 'python_implementation', lambda: None)() == 'PyPy'
-IS_PURE = 'PURE_PYTHON' in os.environ
-
-
-class Acquired:
-    "Marker for explicit acquisition"
-
-
+IS_PURE = int(os.environ.get('PURE_PYTHON', '0'))
+CAPI = not (IS_PYPY or IS_PURE)
+Acquired = "<Special Object Used to Force Acquisition>"
 _NOT_FOUND = object()  # marker
 
 ###
@@ -46,19 +41,24 @@ def _apply_filter(predicate, inst, name, result, extra, orig):
     return predicate(orig, inst, name, result, extra)
 
 
-def _rebound_method(method, wrapper):
-    """Returns a version of the method with self bound to `wrapper`"""
-    if isinstance(method, types.MethodType):
-        method = types.MethodType(method.__func__, wrapper)
-    return method
+if sys.version_info < (3,):  # pragma: PY2
+    PY2 = True
+    import copy_reg
 
+    def _rebound_method(method, wrapper):
+        """Returns a version of the method with self bound to `wrapper`"""
+        if isinstance(method, types.MethodType):
+            method = types.MethodType(method.im_func, wrapper, method.im_class)
+        return method
+else:  # pragma: PY3
+    PY2 = False
+    import copyreg as copy_reg
 
-def _reraise(tp, value, tb=None):
-    if value is None:
-        value = tp()
-    if value.__traceback__ is not tb:
-        raise value.with_traceback(tb)
-    raise value
+    def _rebound_method(method, wrapper):
+        """Returns a version of the method with self bound to `wrapper`"""
+        if isinstance(method, types.MethodType):
+            method = types.MethodType(method.__func__, wrapper)
+        return method
 
 ###
 # Wrapper object protocol, mostly ported from C directly
@@ -306,6 +306,16 @@ def _Wrapper_findattr(wrapper, name,
     raise AttributeError(orig_name)
 
 
+def _Wrapper_fetch(self, name, default=AttributeError):
+    try:
+        return _Wrapper_findattr(self, name, None, None, None, True,
+                                     type(self)._IS_IMPLICIT, False, False)
+    except AttributeError:
+        if type(default) is type and issubclass(default, Exception):
+            raise default(name)
+        return default
+
+
 _NOT_GIVEN = object()  # marker
 _OGA = object.__getattribute__
 
@@ -361,7 +371,17 @@ class _Wrapper(ExtensionClass.Base):
             # so that if it has methods that use `object.__getattribute__`
             # they still work. Note that because we have slots,
             # we won't interfere with the contents of that dict.
-            object.__setattr__(inst, '__dict__', obj.__dict__)
+            od = obj.__dict__
+            if not isinstance(od, dict):
+                # Python 3 refuses to set ``__dict__`` to a non dict
+                # thus, convert
+                # Note: later changes to ``od`` will not be
+                # reflected by the wrapper. But, it is rare
+                # that ``od`` is not a dict (usually for class objects)
+                # and wrappers are transient entities. Thus, the
+                # risk should not be too high.
+                od = dict(od)
+            object.__setattr__(inst, '__dict__', od)
         return inst
 
     def __init__(self, obj, container):
@@ -412,12 +432,14 @@ class _Wrapper(ExtensionClass.Base):
         # the wrapped __of__ method because it gets here via the
         # descriptor code path)...
         wrapper = self._obj.__of__(parent)
-        if (not isinstance(wrapper, _Wrapper) or
-                not isinstance(wrapper._container, _Wrapper)):
+        if not isinstance(wrapper, _Wrapper):
             return wrapper
         # but the returned wrapper should be based on this object's
         # wrapping chain
         wrapper._obj = self
+
+        if not isinstance(wrapper._container, _Wrapper):
+            return wrapper
 
         while (isinstance(wrapper._obj, _Wrapper) and
                (wrapper._obj._container is wrapper._container._obj)):
@@ -486,11 +508,10 @@ class _Wrapper(ExtensionClass.Base):
     # define __cmp__ the same way, and redirect the rich comparison operators
     # to it. (Note that these attributes are also hardcoded in getattribute)
     def __cmp__(self, other):
-        aq_self = self._obj
-        if hasattr(type(aq_self), '__cmp__'):
-            return _rebound_method(aq_self.__cmp__, self)(other)
-
         my_base = aq_base(self)
+        cmp = getattr(type(my_base), "__cmp__", None)
+        if cmp is not None:
+            return cmp(self, other)
         other_base = aq_base(other)
         if my_base is other_base:
             return 0
@@ -514,49 +535,33 @@ class _Wrapper(ExtensionClass.Base):
     def __ge__(self, other):
         return self.__cmp__(other) >= 0
 
-    # Special methods looked up by the type of self._obj,
-    # but which must have the wrapper as self when called
+    # Special methods:
+    # make implicitly called `obj.__method__`
+    # behave the same way as when emplicitly called
 
     def __nonzero__(self):
-        aq_self = self._obj
-        type_aq_self = type(aq_self)
-        nonzero = getattr(type_aq_self, '__nonzero__', None)
+        nonzero = _Wrapper_fetch(self, '__nonzero__', None)
         if nonzero is None:
-            nonzero = getattr(type_aq_self, '__bool__', None)
+            # Py3 bool?
+            nonzero = _Wrapper_fetch(self, '__bool__', None)
         if nonzero is None:
             # a len?
-            nonzero = getattr(type_aq_self, '__len__', None)
+            nonzero = _Wrapper_fetch(self, '__len__', None)
         if nonzero:
-            return bool(nonzero(self))  # Py3 is strict about the return type
+            return bool(nonzero())  # Py3 is strict about the return type
         # If nothing was defined, then it's true
         return True
     __bool__ = __nonzero__
 
     def __unicode__(self):
-        f = getattr(self.aq_self, '__unicode__',
-                    getattr(self.aq_self, '__str__', object.__str__))
-        return _rebound_method(f, self)()
-
-    def __repr__(self):
-        aq_self = self._obj
-        try:
-            return _rebound_method(aq_self.__repr__, self)()
-        except (AttributeError, TypeError):
-            return repr(aq_self)
-
-    def __str__(self):
-        aq_self = self._obj
-        try:
-            return _rebound_method(aq_self.__str__, self)()
-        except (AttributeError, TypeError):
-            return str(aq_self)
+        f = _Wrapper_fetch(self, '__unicode__', None)
+        if f is None:
+            f = _Wrapper_fetch(self, '__str__')
+        return f()
 
     def __bytes__(self):
-        aq_self = self._obj
-        try:
-            return _rebound_method(aq_self.__bytes__, self)()
-        except (AttributeError, TypeError):
-            return bytes(aq_self)
+        return _Wrapper_fetch(self, '__bytes__', TypeError)()
+
 
     __binary_special_methods__ = [
         # general numeric
@@ -619,6 +624,11 @@ class _Wrapper(ExtensionClass.Base):
     ]
 
     __unary_special_methods__ = [
+        # misc
+        '__repr__',
+        # requres ``AttributeError`` --> ``TypeError`` for PY3
+        # '__bytes__',
+        '__str__',
         # arithmetic
         '__neg__',
         '__pos__',
@@ -640,22 +650,12 @@ class _Wrapper(ExtensionClass.Base):
         # '__str__',
     ]
 
-    for _name in __binary_special_methods__:
+    for _name in __unary_special_methods__ + __binary_special_methods__:
         def _make_op(_name):
-            def op(self, other):
-                aq_self = self._obj
-                return getattr(type(aq_self), _name)(self, other)
+            def op(self, *args):
+                return _Wrapper_fetch(self, _name)(*args)
             return op
         locals()[_name] = _make_op(_name)
-
-    for _name in __unary_special_methods__:
-        def _make_op(_name):
-            def op(self):
-                aq_self = self._obj
-                return getattr(type(aq_self), _name)(self)
-            return op
-        locals()[_name] = _make_op(_name)
-
     del _make_op
     del _name
 
@@ -663,20 +663,15 @@ class _Wrapper(ExtensionClass.Base):
 
     def __len__(self):
         # if len is missing, it should raise TypeError
-        # (AttributeError breaks list conversion if it is raised)
-        try:
-            l = getattr(type(self._obj), '__len__')
-        except AttributeError:
-            raise TypeError('object has no len()')
-        else:
-            return l(self)
+        # (AttributeError is acceptable under Py2, but Py3
+        # breaks list conversion if AttributeError is raised)
+        return _Wrapper_fetch(self, '__len__', TypeError)()
 
     def __iter__(self):
-        # For things that provide either __iter__ or just __getitem__,
-        # we need to be sure that the wrapper is provided as self
-        if hasattr(self._obj, '__iter__'):
-            return _rebound_method(self._obj.__iter__, self)()
-        if hasattr(self._obj, '__getitem__'):
+        it = _Wrapper_fetch(self, '__iter__', None)
+        if it is not None:
+            return it()
+        if hasattr(self, '__getitem__'):
             # Unfortunately we cannot simply call iter(self._obj)
             # and rebind im_self like we do above: the Python runtime
             # complains:
@@ -693,37 +688,30 @@ class _Wrapper(ExtensionClass.Base):
             it = WrapperIter(self)
             return iter(it)
 
-        return iter(self._obj)
+        raise TypeError("__iter__")
 
     def __contains__(self, item):
         # First, if the type of the object defines __contains__ then
         # use it
-        aq_self = self._obj
-        aq_contains = getattr(type(aq_self), '__contains__', None)
+        aq_contains = _Wrapper_fetch(self, '__contains__', None)
         if aq_contains:
-            return aq_contains(self, item)
+            return aq_contains(item)
         # Next, we should attempt to iterate like the interpreter;
         # but the C code doesn't do this, so we don't either.
         # return item in iter(self)
         raise AttributeError('__contains__')
 
     def __setitem__(self, key, value):
-        aq_self = self._obj
-        try:
-            setter = type(aq_self).__setitem__
-        except AttributeError:
-            raise AttributeError("__setitem__")  # doctests care about the name
-        else:
-            setter(self, key, value)
+        setter = _Wrapper_fetch(self, "__setitem__")
+        setter(key, value)
 
     def __getitem__(self, key):
-        aq_self = self._obj
-        try:
-            getter = type(aq_self).__getitem__
-        except AttributeError:
-            raise AttributeError("__getitem__")  # doctests care about the name
-        else:
-            return getter(self, key)
+        getter = _Wrapper_fetch(self , '__getitem__')
+        return getter(key)
+
+    if PY2:
+        def __getslice__(self, start, end):
+            return _Wrapper_fetch(self, '__getslice__')(start, end)
 
     def __call__(self, *args, **kwargs):
         try:
@@ -758,15 +746,11 @@ class _Acquirer(ExtensionClass.Base):
 
     def __getattribute__(self, name):
         try:
-            return super().__getattribute__(name)
-        except AttributeError:
+            return super(_Acquirer, self).__getattribute__(name)
+        except AttributeError as exc:
             # the doctests have very specific error message
-            # requirements (but at least we can preserve the traceback)
-            _, _, tb = sys.exc_info()
-            try:
-                _reraise(AttributeError, AttributeError(name), tb)
-            finally:
-                del tb
+            exc.args = AttributeError(name).args
+            raise
 
     def __of__(self, context):
         return type(self)._Wrapper(self, context)
@@ -924,7 +908,7 @@ def aq_inContextOf(self, o, inner=True):
     return False
 
 
-if not (IS_PYPY or IS_PURE):  # pragma: no cover
+if CAPI:  # pragma: no cover
     # Make sure we can import the C extension of our dependency.
     from ExtensionClass import _ExtensionClass  # NOQA
     from ._Acquisition import *  # NOQA