+++ src/Acquisition/tests.py
@@ -34,11 +34,14 @@ from Acquisition import (  # NOQA
     aq_self,
     Explicit,
     Implicit,
+    CAPI,
     IS_PYPY,
     IS_PURE,
+    _Wrapper,
 )
 
 
+<<<<<<< HEAD
 def unicode(self):
     # For test purposes, redirect the unicode
     # to the type of the object, just like Py2 did
@@ -47,6 +50,23 @@ def unicode(self):
     except AttributeError:
         return type(self).__str__(self)
 
+=======
+    def unicode(self):
+        # For test purposes, redirect the unicode
+        # to the type of the object, just like Py2 did
+        try:
+            return type(self).__unicode__(self)
+        except AttributeError as e:
+            return type(self).__str__(self)
+
+    def cmp(x, y):
+        return 0 if x == y else -1 if x < y else 1
+
+    long = int
+else:
+    PY2 = True
+    PY3 = False
+>>>>>>> master
 
 if 'Acquisition._Acquisition' not in sys.modules:
     CAPI = False
@@ -285,7 +305,7 @@ class TestStory(unittest.TestCase):
         # - The object found, and
         # - The extra data passed to 'aq_acquire'.
 
-        # If the filter returns a true object that the object found is
+        # If the filter returns a true object then the object found is
         # returned, otherwise, the acquisition search continues.
         # For example, in:
 
@@ -1747,7 +1767,7 @@ def test_container_proxying():
     ...     def __getitem__(self, key):
     ...         if isinstance(key, slice):
     ...             print('slicing...')
-    ...             return (key.start,key.stop)
+    ...             return (key.start, key.stop, key.step)
     ...         print('getitem', key)
     ...         if key == 4:
     ...             raise IndexError
@@ -1760,7 +1780,7 @@ def test_container_proxying():
     ...         return iter((42,))
     ...     def __getslice__(self, start, end):
     ...         print('slicing...')
-    ...         return (start, end)
+    ...         return (start, end, None)
 
     The naked class behaves like this:
 
@@ -1776,8 +1796,22 @@ def test_container_proxying():
     [42]
     >>> c[5:10]
     slicing...
+<<<<<<< HEAD
     (5, 10)
     >>> c[5:] == (5, None)
+=======
+    (5, 10, None)
+    >>> c[5:] == (5, sys.maxsize if PY2 else None, None)
+    slicing...
+    True
+    >>> c[:10] == (0 if PY2 else None, 10, None)
+    slicing...
+    True
+    >>> c[5:10:5] == (5, 10, 5)
+    slicing...
+    True
+    >>> c[::] == (None, None, None)
+>>>>>>> master
     slicing...
     True
 
@@ -1799,8 +1833,22 @@ def test_container_proxying():
     [42]
     >>> i.c[5:10]
     slicing...
+<<<<<<< HEAD
     (5, 10)
     >>> i.c[5:] == (5, None)
+=======
+    (5, 10, None)
+    >>> i.c[5:] == (5, sys.maxsize if PY2 else None, None)
+    slicing...
+    True
+    >>> i.c[:10] == (0 if PY2 else None, 10, None)
+    slicing...
+    True
+    >>> i.c[5:10:5] == (5, 10, 5)
+    slicing...
+    True
+    >>> i.c[::] == (None, None, None)
+>>>>>>> master
     slicing...
     True
 
@@ -1813,7 +1861,7 @@ def test_container_proxying():
     ...     def __getitem__(self, key):
     ...         if isinstance(key, slice):
     ...             print('slicing...')
-    ...             return (key.start,key.stop)
+    ...             return (key.start, key.stop, key.step)
     ...         print('getitem', key)
     ...         if key == 4:
     ...             raise IndexError
@@ -1826,7 +1874,7 @@ def test_container_proxying():
     ...         return iter((42,))
     ...     def __getslice__(self, start, end):
     ...         print('slicing...')
-    ...         return (start, end)
+    ...         return (start, end, None)
 
     The naked class behaves like this:
 
@@ -1842,8 +1890,22 @@ def test_container_proxying():
     [42]
     >>> c[5:10]
     slicing...
+<<<<<<< HEAD
     (5, 10)
     >>> c[5:] == (5, None)
+=======
+    (5, 10, None)
+    >>> c[5:] == (5, sys.maxsize if PY2 else None, None)
+    slicing...
+    True
+    >>> c[:10] == (0 if PY2 else None, 10, None)
+    slicing...
+    True
+    >>> c[5:10:5] == (5, 10, 5)
+    slicing...
+    True
+    >>> c[::] == (None, None, None)
+>>>>>>> master
     slicing...
     True
 
@@ -1865,8 +1927,22 @@ def test_container_proxying():
     [42]
     >>> i.c[5:10]
     slicing...
+<<<<<<< HEAD
     (5, 10)
     >>> i.c[5:] == (5, None)
+=======
+    (5, 10, None)
+    >>> i.c[5:] == (5, sys.maxsize if PY2 else None, None)
+    slicing...
+    True
+    >>> i.c[:10] == (0 if PY2 else None, 10, None)
+    slicing...
+    True
+    >>> i.c[5:10:5] == (5, 10, 5)
+    slicing...
+    True
+    >>> i.c[::] == (None, None, None)
+>>>>>>> master
     slicing...
     True
 
@@ -2271,6 +2347,32 @@ class TestBugs(unittest.TestCase):
         except AttributeError:
             raise
 
+    def test_wrapped_attr(self):
+        top = I("")
+        top.f = I("f")
+        f = top.f
+        i = I("i")
+        i.f = f
+        self.assertIs(i.f.aq_self, f)
+        self.assertIs(i.f.aq_parent, i)
+
+    def test__cmp__(self):
+        class CmpOrdered(I):
+            def __cmp__(self, other):
+                return cmp(self.id, other.id)
+        Ordered = CmpOrdered
+        top = Ordered("")
+        top.o = Ordered("")
+        o = top.o
+        self.assertEqual(o.o.__cmp__(top), 0)
+        self.assertEqual(o.o.__cmp__(Ordered("1")), -1)
+
+    def test_bool(self):
+        top = I("")
+        top.o = I("")
+        o = top.o
+        self.assertTrue(bool(o.o))
+
 
 class TestSpecialNames(unittest.TestCase):
 
@@ -2467,6 +2569,17 @@ class TestWrapper(unittest.TestCase):
 
         with self.assertRaises(TypeError):
             bytes(wrapper)
+
+    def test_aq_acquire_with_class_and_filter(self):
+        class C(object):
+            a = 1
+        allow = lambda *args: True
+        deny = lambda *args: False
+        self.assertEqual(aq_acquire(C, "a", allow), C.a)
+        with self.assertRaises(AttributeError):
+            aq_acquire(C, "a", deny)
+        with self.assertRaises(AttributeError):
+            aq_acquire(C, "b")
 
 
 class TestOf(unittest.TestCase):
@@ -3199,18 +3312,15 @@ class TestProxying(unittest.TestCase):
 
         self.assertEqual(base.derived(1, k=2), (42, 1, 2))
 
-        if not IS_PYPY:
-            # XXX: This test causes certain versions
-            # of PyPy to segfault (at least 2.6.0-alpha1)
-            class NotCallable(base_class):
-                pass
+        class NotCallable(base_class):
+            pass
 
-            base.derived = NotCallable()
-            try:
-                base.derived()
-                self.fail("Not callable")
-            except (TypeError, AttributeError):
-                pass
+        base.derived = NotCallable()
+        try:
+            base.derived()
+            self.fail("Not callable")
+        except (TypeError, AttributeError):
+            pass
 
     def test_implicit_proxy_call(self):
         self._check_call()
@@ -3333,13 +3443,18 @@ class TestProxying(unittest.TestCase):
 
 class TestCompilation(unittest.TestCase):
 
-    def test_compile(self):
-        if IS_PYPY or IS_PURE:
-            with self.assertRaises((AttributeError, ImportError)):
-                from Acquisition import _Acquisition
-        else:
+    def test_compilation(self):
+        self.assertEqual(CAPI, not (IS_PYPY or IS_PURE))
+        try:
             from Acquisition import _Acquisition
+            cExplicit = _Acquisition.Explicit
+        except ImportError:
+            cExplicit = None  # PyPy never has a C module.
+        if CAPI:  # pragma: no cover
             self.assertTrue(hasattr(_Acquisition, 'AcquisitionCAPI'))
+            self.assertEqual(Explicit, cExplicit)
+        else:
+            self.assertNotEqual(Explicit, cExplicit)
 
 
 def test_suite():